+++ odoo/local-src/connector_esb/components/__init__.py
@@ -9,4 +9,3 @@ from . import importer
 from . import exporter
 from . import events
 from . import webservice_message
-from . import adapters