+++ odoo/local-src/connector_esb/models/sale_order/__init__.py
@@ -1,3 +1,4 @@
+from . import adapter
 from . import common
 from . import exporter
 from . import listener